small checkbox refactor
